Split app into two classes to separate app driver and app following review feedback #1619
